Extract GetDescendantsAcrossDeclarations
